requirements.txt
transformers>=4.51.1
torch>=2.0.0
torchvision>=0.15.0
accelerate>=1.10.0
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
pandas>=2.2.2
pillow>=11.0.0
numpy<2
mistralai>=1.0.0
python-dotenv>=1.0.0
faster-whisper==1.0.3
requests>=2.32.0
tqdm>=4.66
rich>=13.9
python-dateutil>=2.9
streamlit
moviepy
pyyaml
fastapi
uvicorn